modified:   Content/TiledProject/OverWorld1.tmx 	modified:   Engine/ECSystem/Level.java 	modified:   Engine/Graphics/Tile/ShadowLayer.java 	modified:   Engine/Graphics/Tile/TileManager.java 	new file:   Engine/Graphics/Tile/TilemapEntities.java 	new file:   Engine/Graphics/Tile/parseEntity.java 	modified:   Gameplay/Enemies/Enemy.java 	modified:   Gameplay/Enemies/Units/GreenKnight.java 	modified:   Gameplay/Interactives/Blocks/Rock.java 	modified:   Gameplay/Levels/TestRoom.java 	modified:   Gameplay/Levels/TestRoom2.java

+++ Gameplay/Enemies/Units/GreenKnight.java
@@ -16,7 +16,7 @@ public class GreenKnight extends Enemy{
     protected Spritesheet sprite=new Spritesheet("Content/Animations/gknight.png", 16,28);
     
 
-    public GreenKnight(Vector2D<Float> position) {
+    public GreenKnight(Vector2D<Integer> position) {
         super(position);
         setPseudoPosition(25f, 50f);
         SetScale(size);
